Correct an error of omission in the implementation of the truncation operation on POSIX shared memory objects and tmpfs.  Previously, neither of these modules correctly handled the case in which the new size of the object or file was not a multiple of the page size.  Specifically, they did not handle partial page truncation of data stored on swap.  As a result, stale data might later be returned to an application.
Interestingly, a data inconsistency was less likely to occur under tmpfs
than POSIX shared memory objects.  The reason being that a different mistake
by the tmpfs truncation operation helped avoid a data inconsistency.  If the
data was still resident in memory in a PG_CACHED page, then the tmpfs
truncation operation would reactivate that page, zero the truncated portion,
and leave the page pinned in memory.  More precisely, the benevolent error
was that the truncation operation didn't add the reactivated page to any of
the paging queues, effectively pinning the page.  This page would remain
pinned until the file was destroyed or the page was read or written.  With
this change, the page is now added to the inactive queue.

Discussed with:	jhb
Reviewed by:	kib (an earlier version)
MFC after:	3 weeks


git-svn-id: svn://svn.freebsd.org/base/head@229821 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f