At the risk of reducing source compatibility with old NetBSD and Sprite: o) Get rid of some unused macros related to features we don't intend to    provide. o) Get rid of macro definitions for MIPS-I CPUs.  We are not likely to    support anything that predartes MIPS-III. o) Respell MIPS3_* macros as MIPS_*, which is how most of them were being    used already. o) Eliminate a duplicate and mostly-unused set of exception vector macros.
There's still considerable duplication and lots more obsolete in our headers,
but this reduces one of the larger files to a size where one could reckon
about the correctness of its contents with a mere few hours of contemplation.

There is, of course, a question of whether we need definitions for fields,
registers and configurations that we are unlikely to ever use or implement,
even if they're not obsolete since 1991.  FreeBSD is not a processor
reference manual, and things that aren't used may be wrong, or may be
duplicated because nobody could possibly actually know whether they're
already defined.


git-svn-id: svn://svn.freebsd.org/base/head@232615 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f